pyproject.toml
[tool.poetry]
name = "fast-gpt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
jinja2 = "^3.1.4"
openai = "^1.33.0"
python-dotenv = "^1.0.1"
markdown = "^3.6"
pygments = "^2.18.0"
anthropic = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
